Add 12V rail for N64 + silkscreen updates
Add a 12V rail for the N64 slot. This is not currently implemented, but as the six-slot adapter is the hardest of the 2 boards to upgrade having it now will save some effort for anyone wanting it in the future. It also makes it easier to externally supply 12V by attaching clips to the 12V and GND header pins.

Also improved the silkscreen a bit to match the main PCB.